refactor: extract method_source_walker as shared leaf below codegen and queries
Move `find_all_sends_in_source`, `find_all_references_in_source`,
`ReceiverKind`, `SendHit`, `ReferenceHit`, and `selector_span` into a
new `method_source_walker` crate-root module that sits below both the
Language Service (`queries/`) and Code Generation (`codegen/`) contexts.

Previously `codegen/core_erlang/gen_server/methods.rs` imported
`crate::queries::all_sends_query` and
`crate::queries::references_to_query` directly, violating the
architecture rule "codegen MUST NOT import queries" (Language Service
layer). The fix applies the Shared-Leaf-Module Pattern
(architecture-principles.md §6): the shared logic is extracted to a
leaf module reachable from both sides without a cross-layer dependency.

- `queries/all_sends_query` and `queries/references_to_query` now
  re-export the moved symbols so existing callers are unchanged.
- `queries/mod.rs` re-exports `selector_span` from the new module so
  the four query modules using `super::selector_span` need no edits.
- `codegen` imports from `crate::method_source_walker` directly,
  removing the forbidden `queries` dependency.

All 5680 unit, 3291 BUnit, 233 stdlib, and runtime tests pass.
